# config/openapi.yaml
openapi: 3.0.3
info:
  title: DocVault Document Archiving System API
  description: |
    REST API for the DocVault Document Archiving System - a comprehensive solution for document storage,
    OCR processing, and intelligent search capabilities.

    ## Features
    - Secure document upload and storage
    - OCR text extraction
    - Full-text search with Meilisearch
    - Role-based access control (RBAC)
    - Document categorization and tagging
    - Audit logging

    ## Authentication
    All API endpoints (except /api/auth/*) require JWT authentication.
    Include the token in the Authorization header: `Authorization: Bearer {token}`
  version: 1.0.0
  contact:
    name: DocVault API Support
    email: support@docvault.local
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://docvault.example.com/api
    description: Production server
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
          format: int64
        filePath:
          type: string
        ocrText:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
        processingStatus:
          type: string
          enum: [uploaded, queued, processing, completed, failed]
          description: Current processing status of the document
        progress:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
          description: Processing progress percentage (0-100). Only present during processing.
          example: 75
        currentOperation:
          type: string
          nullable: true
          description: Description of current processing operation. Only present during processing.
          example: "Performing OCR on page 5/10"
        thumbnailPath:
          type: string
          nullable: true
        categoryId:
          type: string
          format: uuid
          nullable: true
        tags:
          type: array
          items:
            type: string
        uploadedBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
    DocumentList:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
paths:
...
  /documents:
    get:
      tags:
        - Documents
      summary: List all documents
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: category
          in: query
          schema:
            type: string
            format: uuid
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt, updatedAt, filename, fileSize]
            default: createdAt
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentList'

    post:
      tags:
        - Documents
      summary: Create a new document (metadata only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - filename
                - mimeType
                - fileSize
              properties:
                filename:
                  type: string
                mimeType:
                  type: string
                fileSize:
                  type: integer
                metadata:
                  type: object
      responses:
        '201':
          description: Document created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
...